remove CommandError from api v2
