+++ b2sdk/exception.py
@@ -172,21 +172,6 @@ class ClockSkew(B2HttpCallbackPostRequestException):
             )
 
 
-class CommandError(B2Error):
-    """
-    b2 command error (user caused).  Accepts exactly one argument: message.
-
-    We expect users of shell scripts will parse our ``__str__`` output.
-    """
-
-    def __init__(self, message):
-        super(CommandError, self).__init__()
-        self.message = message
-
-    def __str__(self):
-        return self.message
-
-
 class Conflict(B2SimpleError):
     pass
